Add fzf aliaes file, and upgrade installing way.

+++ install.sh
@@ -1,25 +1,46 @@
 #!/bin/env bash
 
-echo "ln  .bash_aliases ~/.bash_aliases"
-ln  ./.bash_aliases ~/.bash_aliases
+if [ ! -f ~/.bash_aliases ]; then
+	echo "ln  .bash_aliases ~/.bash_aliases"
+	ln  ./.bash_aliases ~/.bash_aliases
+fi
 
-echo "ln  .gitconfig ~/.gitconfig"
-ln  ./.gitconfig ~/.gitconfig
+if [ ! -f ~/.gitconfig ]; then
+	echo "ln  .gitconfig ~/.gitconfig"
+	ln  ./.gitconfig ~/.gitconfig
+fi
 
-echo "ln  .gitignore ~/.gitignore"
-ln  ./.gitignore ~/.gitignore
+if [ ! -f ~/.gitignore ]; then
+	echo "ln  .gitignore ~/.gitignore"
+	ln  ./.gitignore ~/.gitignore
+fi
 
-echo "ln  .Xmodmap ~/.Xmodmap"
-ln  ./.Xmodmap ~/.Xmodmap
+if [ ! -f ~/.Xmodmap ]; then
+	echo "ln  .Xmodmap ~/.Xmodmap"
+	ln  ./.Xmodmap ~/.Xmodmap
+fi
 
-echo "ln  .tmux.conf ~/.tmux.conf"
-ln  ./.tmux.conf ~/.tmux.conf
+if [ ! -f ~/.tmux.conf ]; then
+	echo "ln  .tmux.conf ~/.tmux.conf"
+	ln  ./.tmux.conf ~/.tmux.conf
+fi
 
-echo "ln  .vimperatorrc ~/.vimperatorrc"
-ln  ./.vimperatorrc ~/.vimperatorrc
+if [ ! -f ~/.vimperatorrc ]; then
+	echo "ln  .vimperatorrc ~/.vimperatorrc"
+	ln  ./.vimperatorrc ~/.vimperatorrc
+fi
 
-echo "ln  .Xresources ~/.Xresources"
-ln  ./.Xresources ~/.Xresources
+if [ ! -f ~/.Xresources ]; then
+	echo "ln  .Xresources ~/.Xresources"
+	ln  ./.Xresources ~/.Xresources
+fi
 
-echo "ln  .Xdefaults ~/.Xdefaults"
-ln  ./.Xdefaults ~/.Xdefaults
+if [ ! -f ~/.Xdefaults ]; then
+	echo "ln  .Xdefaults ~/.Xdefaults"
+	ln  ./.Xdefaults ~/.Xdefaults
+fi
+
+if [ ! -f ~/.fzf_aliases ]; then
+	echo "ln  .fzf_aliases ~/.fzf_aliases"
+	ln  ./.fzf_aliases ~/.fzf_aliases
+fi
